// db-consumer/src/interfaces/Error.ts
export interface AppError {
    err: Errors;
}

export enum Errors {
    RABBITMQ_RECEIVED_INVALID_TOPIC = "RABBITMQ_RECEIVED_INVALID_TOPIC",
    RABBITMQ_RECEIVED_NULL_MESSAGE = "RABBITMQ_RECEIVED_NULL_MESSAGE",

    RAW_VALIDATION_FAILED = "RAW_VALIDATION_FAILED",
    PARSED_VALIDATION_FAILED = "PARSED_VALIDATION_FAILED",

    PARSED_MALFORMED_JSON = "PARSED_MALFORMED_JSON",
    RAW_MALFORMED_JSON = "RAW_MALFORMED_JSON",

    RAW_DB_SAVE_FAILED = "RAW_DB_SAVE_FAILED",
    PARSED_DB_SAVE_FAILED = "PARSED_DB_SAVE_FAILED",

    RAW_DB_CHECK_FAILED = "RAW_DB_CHECK_FAILED",
    PARSED_DB_CHECK_FAILED = "PARSED_DB_CHECK_FAILED",
    RAW_DATA_NOT_FOUND = "RAW_DATA_NOT_FOUND",

    GEOLOCATION_API_FAILED = "GEOLOCATION_API_FAILED",

    FB_PARSING_FAILED = "FB_PARSING_FAILED",

    RECEIVED_INVALID_SCRAPER_TYPE = "RECEIVED_INVALID_SCRAPER_TYPE",

    ERROR_FETCHING_RAW_DATA = "ERROR_FETCHING_RAW_DATA"
}